remove vestigial files
